:sparkles: Add project functionality
add the add project functionality
